% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_theme_ft.R
\name{load_theme_ft}
\alias{load_theme_ft}
\title{A function to plot all the palettes in gchartcolour for colour picking}
\usage{
load_theme_ft()
}
\description{
This function loads a basic FT theme
}
\examples{
load_theme_ft()
}
