% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_theme_ft.R
\docType{data}
\name{theme_ft}
\alias{theme_ft}
\title{#' A function to plot all the palettes in gchartcolour for colour picking
#'
#' This function loads a basic FT theme
#' @export
#' @examples
#' load_theme_ft()}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 68.}
\usage{
theme_ft
}
\description{

}
\keyword{datasets}
