% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_theme_ft.R
\docType{data}
\name{theme_ft}
\alias{theme_ft}
\title{A function to add an FT theme to the global environment}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 68.}
\usage{
theme_ft
}
\description{
This function adds an object theme_ft, to the global environment. It is used like this plot + theme_ft()
}
\examples{
load_theme_ft()
}
\keyword{datasets}
